Reorganize files, apply coords and citymapper to dataframe
Also worked out some issues, but so many more came up in the meantime
that it's not worth trying to document. Bottom line, the citymapper
API is not worth the trouble. I don't think the results are accurate.
Will be better to explore Bing's API for routes, and possibly they have
a batch get too. ˆ