Redesign scorecard as JSON for easier maintenance and usage.
Originally, the CSV was a terrible format but I had tunnel vision about
the process. Now, I'll manually edit this JSON but it's in a more
intuitive dictionary-like format. It also allows for multiple values
to have the same point score. I.e., if I get around to scoring home
types, 'Colonial' and 'Rancher' can both be 2.0, instead of having to
rank them or write a separate function.