Have webdriver wait until listing page is fully loaded
Related to a weird issue where the soup object wouldn't have certain
attributes because it grabbed the "Loading..." page. Apparently, my
click wait used to solve this but I took that away because it was
otherwise unnecessarily long. With this fix, the page looks for the
listing details every 500 milliseconds and not a moment longer.