Re-add step to convert all to integer, exclude certain keys
I nix'ed this else clause to coerce everything to integers because it
seemed like a poor catch-all and it was converting my carefully
constructed floats. However, in doing that, it stopped converting
all the many little integers that I didn't want to specifically
enumerate.

So instead, I added that step back in and just excluded the categories
that have floats I wanted to protect.