Condense all uses of `continuous_score()` into single loop
Instead of storing the min/max, weight, etc. for each run of this
function in the code itself, moved the parameters to the scorecard.json
file so they can be edited outside of the code.

The new score2.py reads in either one of the scorecards (regular or
continuous), and will unpack values for the very complex
continuous_score() function each time.