refactor: create class for houses
Previously, initialized a dict with existing keys, which would be updated as it scraped the page.
Now, uses class House (a dict subclass) to store scraped data but also have its own scrape method.
This puts me in a much better position to manage the scraping, cleaning, and exporting of individual
houses or groups of them.