fix (database): replace context manager with HTTP POST request
The context manager in the Cloudant API for accessing a Document was
opening the doc for editing and assuming that an edit was made, thus
incrementing the _rev each time even though no change was made.

This commit replaces that method of getting the remote doc for
comparison with a standard POST request via the Requests session object.

Also:
- refactor: renamed many variables, moved functions to nested functions
in order to reduce variable name shadowing and docstring length