feat (scrape): create webdriver if none passed to scrape
Previously, a webdriver needed to be passed to the home instance's
scrape() method. When not in the context manager for a bulk scrape,
one had to be instantiated manually which was tough to do in a command
line environment, and had to be repeated in each script that I tested
in.

Now, it can accept a webdriver (for when it's part of a bulk scrape), or
it will create one specifically for the single scrape.