fix (clean): removed loggers and added exception handler
Loggers were getting to be too verbose (they were initially to debug
what things weren't getting cleaned in the dic), and their log
statements were referencing things that may not be in a home instance.

Also adds an exception handler to the number parser. Now all actions
in clean.py have exception handlers for key errors, so it should be
safe to run on an empty dictionary as well as a screwed up one without
killing the whole program.