feat: call enriching functions, fix address parser
The address parser was overwriting itself whenever multiple
values corresponded to the same key (e.g. a two-word street
name). This happens because of how the usaddress package
returns tuples with the keys and values swapped, basically.

This commit stores all tuples in a defaultdict, appending
additional values to a list, then joins those lists as a string
at the end.